+++ Test/UnitTests/OldTestSupportIssueTests/Issue012Tests.cs
@@ -45,7 +45,7 @@ namespace Test.UnitTests.OldTestSupportIssueTests
                 var database = factory.Create(_connectionString,
                     new DatabaseModelFactoryOptions(new string[] { }, new string[] { }));
 
-                var handler = new Stage1Comparer(context.Model, nameof(Issue012DbContext));
+                var handler = new Stage1Comparer(context);
 
                 //ATTEMPT
                 var hasErrors = handler.CompareModelToDatabase(database);