+++ ReleaseNotes.md
@@ -10,12 +10,10 @@
   - ValueConverters
   - Alternative keys
   - Support Views
-  - Use IRelationalTypeMappingSource for constants - see EF Core #21731
   - EF Core 5 - TPT
   - EF Core 5 - Owned Type class can now be required  
   - EF Core 5 - property bag
   - EF Core 5 - Same type mapped to different tables ??
-  - EF Core 5 = persistant computed column
   - EF Core 5 - Map entity to query
   - EF Core 5 - IPAddress and PhysicalAddress
   - Change DesignProvider to use EF Core constant instead of strings
@@ -27,4 +25,6 @@
 ## 1.0.0-preview001
 
 - Support for EF Core 5 
+- Used IRelationalTypeMappingSource for constants - see EF Core #21731
 - EF Core 5 - direct many-to-many
+- EF Core 5 = persistent computed column